clase operador con validaciones
cambios en los js y en las vistas y controladores, funcional.

+++ public/scripts/user/script.js
@@ -1,8 +1,10 @@
 const openShadowContainer = document.getElementsByClassName('open-shadow-container')[0]
+const openShadowContainerOperator = document.getElementsByClassName('open-shadow-container-operator')[0]
 const closeShadowContainer = document.getElementsByClassName('close-shadow-container')[0]
 const shadowContainer = document.getElementsByClassName('shadow-container')[0]
 
 openShadowContainer.addEventListener('click', () => {
+    
     document.getElementById('changePasswordField').style.display = 'none';
     document.querySelector('input[name="id"]').value = '';
     document.querySelector('input[name="name"]').value = '';
@@ -15,46 +17,59 @@ openShadowContainer.addEventListener('click', () => {
     document.getElementById('password-div').style.display = 'block'
 })
 
-
-
-
-closeShadowContainer.addEventListener('click', event => {
-  event.preventDefault()
+openShadowContainerOperator.addEventListener('click', () => {
+  
+  document.getElementById('changePasswordField').style.display = 'none';
+  document.querySelector('input[name="id"]').value = '';
+  document.querySelector('input[name="name"]').value = '';
+  document.querySelector('input[name="email"]').value = '';
   shadowContainer.classList.toggle('toggle-shadow-container')
-  document.querySelector('input[name="password"]').value = '';
+  document.getElementById('changePassword').checked = true;
+  document.getElementById('password-div').style.display = 'block'
 })
 
-// operator value
+
 document.getElementsByTagName('table')[0].addEventListener('click', event => {
-  event.preventDefault()
+  event.preventDefault();
 
   if (event.target.getAttribute('value-rol') == 'operator') {
     if (event.target.innerText == 'Borrar') {
-      const form = new FormData()
-      form.append('id', event.target.getAttribute('value-id'))
+      // Mostrar un cuadro de diálogo de confirmación
+      if (confirm('¿Estás seguro de que deseas eliminar este registro?')) {
+        const form = new FormData();
+        form.append('id', event.target.getAttribute('value-id'));
 
-      fetch('/api/deleteOperator/' + event.target.getAttribute('value-id'), {
-        method: 'delete',
-        body: form
-      }).then(res => {
-        return res.text()
-      }).then(res => {
-        window.location.reload()
-        console.log(res)
-      })
+        fetch('/api/deleteOperator/' + event.target.getAttribute('value-id'), {
+          method: 'delete',
+          body: form
+        }).then(res => {
+          return res.text();
+        }).then(res => {
+          window.location.reload();
+          console.log(res);
+        });
+      }
     }
   }
-})
+});
+
+
+
 
 document.getElementsByTagName('table')[0].addEventListener('click', event => {
   event.preventDefault()
 
   if (event.target.getAttribute('value-rol') == 'operator') {
     if (event.target.innerText == 'Modificar') {
-      shadowContainer.classList.toggle('toggle-shadow-container')
+      
+      document.getElementById('changePasswordField').style.display = 'block'
       document.querySelector('input[name="id"]').value = event.target.getAttribute('value-id')
       document.querySelector('input[name="name"]').value = event.target.getAttribute('value-name')
       document.querySelector('input[name="email"]').value = event.target.getAttribute('value-email')
+      document.getElementById('changePassword').checked = false;
+      document.getElementById('password-div').style.display = 'none'
+      shadowContainer.classList.toggle('toggle-shadow-container')
+
     }
   }
 })
